everything int16 or 32

+++ modelestimator/_bw_estimator/_calculate_q_eq/calculate_q_eq.py
@@ -14,7 +14,8 @@ def calculate_q_eq(count_matrix_list, threshold):
     vl, vr, eq = find_eigens(count_matrix_list)
 
     #   Get a first simple estimate of Q using a Jukes-Cantor model
-    dist_samples = np.arange(1, 400, 5)
+    #DTYPECHANGE
+    dist_samples = np.arange(1, 400, 5, dtype=np.int16)
     posterior = comp_posterior_JC(count_matrix_list, dist_samples)   # posterior.shape = (10, 80). Rows are identical to Octave but in different order
     pw = matrix_weight(count_matrix_list, posterior, dist_samples)    
     w = posterior.sum(axis=0)
@@ -29,6 +30,7 @@ def calculate_q_eq(count_matrix_list, threshold):
     while (iterations < max_iterations and difference > threshold):
         iterations += 1
         q_new = simple_estimation(count_matrix_list, q, vl, vr, eq, dist_samples)
+        #DTYPECHANGE
         difference = np.linalg.norm(q_new - q)
         q = q_new
     return q, eq
